Add files via upload
Pandas is a popular open-source data manipulation and analysis library for Python. It provides data structures for efficiently storing and manipulating large datasets and offers a variety of functions for data cleaning, exploration, and analysis. The primary data structures in Pandas are:

DataFrame: A two-dimensional table with labeled axes (rows and columns). It is similar to a spreadsheet or SQL table, and you can think of it as a container for storing and manipulating tabular data.

Series: A one-dimensional labeled array that can hold any data type. It is similar to a column in a DataFrame or a single variable in statistics.

Pandas allows you to perform a wide range of operations on datasets, such as:

Reading and writing data from/to different file formats (CSV, Excel, SQL databases, etc.).
Cleaning and handling missing data.
Filtering and selecting specific subsets of data.
Grouping and aggregating data.
Merging and joining datasets.
Time series analysis.
Plotting data for visualization.
The library is widely used in data science, machine learning, and other fields where working with structured data is a common task. The flexibility and ease of use provided by Pandas make it an essential tool for data analysis in Python.